Upgrade Hibernate 3.3.2.GA to 3.5.6-Final.
Do upgrade changes as per Hibernate's 3.5 upgrade guide. These changes
have not been tested.

pom.xml:
- Upgrade Hibernate dependencies from 3.3.2.GA to 3.5.6-Final.

Permission.java, PermissionTemplate.java, Profile.java, ProfileOverrides.java,
Site.java, TargetInstance.java:
- Change type='text' to type='materialized_clob' for SQL types clob (Oracle) or
  text (MySQL and Postgres), as this is the type that now maps to SQL clob.

HarvestCoordinatorImpl.java, InTrayManager.java, InTrayManagerImpl.java,
ProfileManager.java, TargetInstanceManager.java,
TargetInstanceManagerImpl.java, SiteManager.java, SiteManagerImpl.java,
TargetManager.java, TargetManagerImpl.java, InTrayDAO.java, ProfileDAO.java,
ProfileDAOImpl.java, SiteDAO.java, SiteDAOImpl.java, TargetDAO.java,
TargetDAOImpl.java, TargetInstanceDAO.java, TargetInstanceDAOImpl.java,
HomeController.java, SitePermissionHandler.java, TransferSeedsController.java,
MockInTrayDAO.java, MockProfileDAO.java, MockSiteDAO.java, MockTargetDAO.java,
MockTargetInstanceDAO.java:
- Change all methods dealing with Projections.rowCount() from int to long, as
  the underlying HQL type has changed. This means method signatures have
  changed from int to long, Object casts are from Integer to Long, and
  calculations and mocking are now long.

+++ src/main/java/org/webcurator/core/notification/InTrayManager.java
@@ -89,7 +89,7 @@ public interface InTrayManager {
      * @param user the user to count the notifications for
      * @return a count of notifications
      */
-    int countNotifications(User user);
+    long countNotifications(User user);
     
     /**
      * gets the List of Tasks appropriate for the user
@@ -104,7 +104,7 @@ public interface InTrayManager {
      * @param user the user to count tasks for
      * @return the count of tasks
      */
-    int countTasks(User user);
+    long countTasks(User user);
     
     /**
      *  generates a Task in the System for all Users within an Agnecy with a particular privilege
@@ -133,7 +133,7 @@ public interface InTrayManager {
      * @param wctResource The related resource.
      * @return The number of tasks.
      */
-    public int countTasks(String messageType, InTrayResource wctResource);
+    public long countTasks(String messageType, InTrayResource wctResource);
     
     
     /**
